fix: do not stream output (#7)
This PR sets the `stream-channels` option to `false` which disables streaming of output.

We are buffering the output so there is no need to stream or have the complexity of a newline delimited parser to deal with it.

This also fixes a problem with IPFS Clusters hosted behind an nginx reverse proxy. Currently if any response is received before the request has finished uploading then nginx doesn't know how to deal with that and terminates the connection. What this means is that if you upload a small file then a big file the request will fail whereas a big file followed by a small file _might_ succeed.